backend/requirements.txt
psycopg2
django
youtube-dl
python-mpd2
djangorestframework
channels
asgi_redis
celery
mutagen
pytaglib
